pyproject.toml
[tool.poetry]
name = "dota2-cast-assist"
version = "0.1.0"
description = "Real-time assistant for broadcasters, integrating Dota 2’s live API and GSI to enhance commentary"
authors = ["Vlad <dota2-cast-assist@scrapingmartians.com>"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11.7"
fastapi = "^0.111.0"
toml = "^0.10.2"
pydantic-settings = "^2.6.0"
google-cloud-firestore = "^2.19.0"
google-cloud-secret-manager = "^2.21.0"
google-cloud-pubsub = "^2.26.1"
google-cloud-storage = "^2.18.2"
pytest-mock = "^3.14.0"
types-requests = "^2.32.0.20241016"
types-toml = "^0.10.8.20240310"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
build = "^1.2.2.post1"
pytest = "^8.3.3"
mypy = "^1.13.0"
tox = "^4.23.2"

[tool.poetry.group.apache-beam.dependencies]
google-apitools = ">=0.5.31,<0.5.32"
apache-beam = {extras = ["gcp"], version = "^2.60.0"}
chardet = "^5.2.0"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "F", "E", "W", "PL"]
ignore = [
    "PLR0912",
    "PLR0913",
    "PLR0914",
    "PLR0915",
    "PLR1704",
    "PLR1722",
    "PLW0108",
    "PLW0120",
    "PLW1510",
    "PLW1514",
]
